Merge pull request to develop

+++ src/CacheInvalidation.Api/UseCases/Handlers/ProductActivedEventHandler.cs
@@ -11,24 +11,33 @@ namespace CacheInvalidation.Api.UseCases.Handlers
         private readonly ICacheDatabase _cacheDatabase;
         private readonly CacheConfig _cacheConfig;
         private readonly ILogger<ProductActivedEventHandler> _logger;
+        private readonly RefreshProductCache _refreshProductCache;
 
-        public ProductActivedEventHandler(ICacheDatabase cacheDatabase, IConfiguration configuration, ILogger<ProductActivedEventHandler> logger)
+        public ProductActivedEventHandler(ICacheDatabase cacheDatabase, IConfiguration configuration, ILogger<ProductActivedEventHandler> logger, RefreshProductCache refreshProductCache)
         {
             this._cacheDatabase = cacheDatabase;
             this._cacheConfig = configuration.GetSection("CacheConfig").Get<CacheConfig>();
             this._logger = logger;
+            this._refreshProductCache = refreshProductCache;
         }
 
-        public Task HandleAsync(ProductActivedEvent notification, CancellationToken cancellationToken = default)
+        public async Task HandleAsync(ProductActivedEvent notification, CancellationToken cancellationToken = default)
         {
-            this._cacheDatabase.RemoverAsync(this._cacheConfig.ProductCacheKey, cancellationToken);
+            if (this._cacheConfig.ItsToRefresh.GetValueOrDefault() == true)
+            {
+                await this._refreshProductCache.ExecuteAsync(cancellationToken);
+            }
+            else
+            {
+                await this._cacheDatabase.RemoverAsync(this._cacheConfig.ProductCacheKey, cancellationToken);
+            }
             var output = new OutboxMessage(notification.Product.Id,
                 nameof(ProductActivedEvent).ToString(),
                 JsonSerializer.Serialize(notification),
                 DateTime.UtcNow,
                 false);
             this._logger.LogInformation($"Product actived event handled: {JsonSerializer.Serialize(output)}");
-            return Task.CompletedTask;
+            
         }
     }
 }
